// supervisorclient/src/app/openapi/model/restartType.ts
/**
 * TaskMaster api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type RestartType = 'ALWAYS' | 'NEVER' | 'ON_FAILURE';

export const RestartType = {
    Always: 'ALWAYS' as RestartType,
    Never: 'NEVER' as RestartType,
    OnFailure: 'ON_FAILURE' as RestartType
};